Added port number mapping
git-svn-id: svn+ssh://software.sandia.gov/svn/private/kratos@181 a2794be9-68be-4fde-8a1b-ce5afca74b0a

+++ getRawPower.c
@@ -62,7 +62,7 @@ int main(int argc, char *argv[])
         // Print results
 	if( timeCollect ) {
         	printf("%ld %ld.%06ld %-4d %5d %5d %7d %7d %7d\n", 
-            		start.tv_sec, now.tv_sec - start.tv_sec, now.tv_usec,
+            		start.tv_sec, now.tv_sec - start.tv_sec, (long int)(now.tv_usec),
             		portNumber, sample.Asamp, sample.Vsamp/16,
             		sample.miliamps,sample.milivolts,sample.miliwatts );
 	} else {
